requirements.txt
streamlit
google-generativeai
requests
beautifulsoup4
tldextract